+++ CleverTap/Editor/AndroidProjectPostProcessor.cs
@@ -228,7 +228,7 @@ namespace CleverTapSDK.Private
 
             if (settings.Environments == null || settings.Environments.Items == null)
             {
-                Debug.LogError("[CTExample] CleverTapSettings - Environments are not configured.");
+                Debug.LogError("[CleverTap] CleverTapSettings - Environments are not configured.");
                 return;
             }
 
@@ -236,13 +236,13 @@ namespace CleverTapSDK.Private
 
             if (environmentCredentials.Count == 0)
             {
-                Debug.LogError("[CTExample] CleverTapSettings - Environments are not configured.");
+                Debug.LogError("[CleverTap] CleverTapSettings - Environments are not configured.");
                 return;
             }
 
             if (!environmentCredentials.TryGetValue(settings.DefaultEnvironment, out CleverTapEnvironmentCredential environmentCredential))
             {
-                Debug.LogError($"[CTExample] CleverTapSettings - Environment is null or not configured for {settings.DefaultEnvironment}");
+                Debug.LogError($"[CleverTap] CleverTapSettings - Environment is null or not configured for {settings.DefaultEnvironment}");
                 return;
             }